Move all setup components into setup folder
