+++ client/extensions/woocommerce/app/dashboard/index.js
@@ -46,10 +46,10 @@ import ManageExternalView from './manage-external-view';
 import ManageNoOrdersView from './manage-no-orders-view';
 import ManageOrdersView from './manage-orders-view';
 import Placeholder from './placeholder';
-import StoreLocationSetupView from './store-location-setup-view';
+import StoreLocationSetupView from './setup/store-location';
 import RequiredPagesSetupView from './required-pages-setup-view';
 import RequiredPluginsInstallView from './required-plugins-install-view';
-import SetupTasksView from './setup-tasks-view';
+import SetupTasksView from './setup';
 import MailChimp from 'woocommerce/app/settings/email/mailchimp/index.js';
 import QuerySettingsGeneral from 'woocommerce/components/query-settings-general';
 import warn from 'lib/warn';
