+++ client/extensions/woocommerce/app/dashboard/setup/notices.js
