+++ client/extensions/woocommerce/app/dashboard/setup/store-location.js
@@ -30,9 +30,9 @@ import {
 import { isCurrentUserEmailVerified } from 'state/current-user/selectors';
 import { isStoreManagementSupportedInCalypsoForCountry } from 'woocommerce/lib/countries';
 import { setSetStoreAddressDuringInitialSetup } from 'woocommerce/state/sites/setup-choices/actions';
-import SetupFooter from './setup-footer';
-import SetupHeader from './setup-header';
-import SetupNotices from './setup-notices';
+import SetupFooter from './footer';
+import SetupHeader from './header';
+import SetupNotices from './notices';
 import { doInitialSetup } from 'woocommerce/state/sites/settings/actions';
 import QueryContactDetailsCache from 'components/data/query-contact-details-cache';
 import QueryLocations from 'woocommerce/components/query-locations';
@@ -218,6 +218,7 @@ class StoreLocationSetupView extends Component {
 		} );
 		const submitDisabled = this.state.isSaving || ! everyRequiredFieldHasAValue;
 
+		/* eslint-disable wpcalypso/jsx-classname-namespace */
 		if ( ! showForm ) {
 			return (
 				<div className="dashboard__placeholder">
@@ -225,12 +226,12 @@ class StoreLocationSetupView extends Component {
 				</div>
 			);
 		}
+		/* eslint-enable wpcalypso/jsx-classname-namespace */
 
 		return (
 			<div>
 				<AddressView
 					address={ this.state.address }
-					className="dashboard__pre-setup-address"
 					isEditable
 					onChange={ this.onChange }
 					showAllLocations
@@ -258,12 +259,12 @@ class StoreLocationSetupView extends Component {
 		const { siteId, translate } = this.props;
 
 		return (
-			<div className="dashboard__setup-wrapper">
+			<div className="setup__wrapper">
 				<SetupNotices />
 				{ this.state.showEmailVerificationDialog && (
 					<VerifyEmailDialog onClose={ this.closeVerifyEmailDialog } />
 				) }
-				<div className="card dashboard__location">
+				<div className="setup__location card">
 					<SetupHeader
 						imageSource={ '/calypso/images/extensions/woocommerce/woocommerce-setup.svg' }
 						imageWidth={ 160 }
